backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.4.4
PyJWT==2.7.0
Werkzeug
python-dotenv==1.0.0
celery==5.3.1
redis==4.6.0
SQLAlchemy==2.0.19